backend/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.0
bcrypt==4.1.1
PyJWT==2.8.0
pytest==8.3.3
httpx==0.27.2
